+++ modules/cloudsim/src/main/java/org/mysim/utils/Parameters.java
@@ -124,7 +124,7 @@ public class Parameters {
     //------------------------------------ Experiment Parameters
     public static VM_SELECTION_TYPE Packing_VM_SELECTION_TYPE = VM_SELECTION_TYPE.RANDOM;
 //    public static VM_SELECTION_TYPE Packing_VM_SELECTION_TYPE = VM_SELECTION_TYPE.BI_FACTOR;
-    public static int R_T_Q_SCHEDULING_INTERVAL = 100;
+    public static double R_T_Q_SCHEDULING_INTERVAL = 100;
     public static int MONITORING_INTERVAL = 50;
     public static double VM_THRESH_HOLD_FOR_SHUTDOWN = 60;
     public static final int CHECK_FINISHED_STATUS_DELAY = 200;
@@ -150,7 +150,7 @@ public class Parameters {
         workflowsDirectory = workflowsPath;
     }
     public static void init(String workflowsPath, double dFactor, double bFactor, double arrivalRate,
-                            int schedulingInterval, int monitoringInterval, double vmShutdown, Boolean degradationEnable) {
+                            double schedulingInterval, int monitoringInterval, double vmShutdown, Boolean degradationEnable) {
         workflowsDirectory = workflowsPath;
         ALPHA_DEADLINE_FACTOR = dFactor;
         BETA_BUDGET_FACTOR = bFactor;
